+++ core/src/server/grpc_impl/GrpcRequestHandler.cpp
@@ -15,32 +15,12 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "server/grpc_impl/GrpcRequestHandler.h"
-
-#include <src/server/Config.h>
-
 #include <memory>
 #include <unordered_map>
 #include <vector>
 
-#include "server/grpc_impl/GrpcRequestScheduler.h"
-#include "server/grpc_impl/request/CmdRequest.h"
-#include "server/grpc_impl/request/CountTableRequest.h"
-#include "server/grpc_impl/request/CreateIndexRequest.h"
-#include "server/grpc_impl/request/CreatePartitionRequest.h"
-#include "server/grpc_impl/request/CreateTableRequest.h"
-#include "server/grpc_impl/request/DeleteByDateRequest.h"
-#include "server/grpc_impl/request/DescribeIndexRequest.h"
-#include "server/grpc_impl/request/DescribeTableRequest.h"
-#include "server/grpc_impl/request/DropIndexRequest.h"
-#include "server/grpc_impl/request/DropPartitionRequest.h"
-#include "server/grpc_impl/request/DropTableRequest.h"
-#include "server/grpc_impl/request/HasTableRequest.h"
-#include "server/grpc_impl/request/InsertRequest.h"
-#include "server/grpc_impl/request/PreloadTableRequest.h"
-#include "server/grpc_impl/request/SearchRequest.h"
-#include "server/grpc_impl/request/ShowPartitionsRequest.h"
-#include "server/grpc_impl/request/ShowTablesRequest.h"
+#include "server/Config.h"
+#include "server/grpc_impl/GrpcRequestHandler.h"
 #include "tracing/TextMapCarrier.h"
 #include "tracing/TracerUtil.h"
 #include "utils/TimeRecorder.h"
@@ -48,6 +28,48 @@
 namespace milvus {
 namespace server {
 namespace grpc {
+
+::milvus::grpc::ErrorCode
+ErrorMap(ErrorCode code) {
+    static const std::map<ErrorCode, ::milvus::grpc::ErrorCode> code_map = {
+        {SERVER_UNEXPECTED_ERROR, ::milvus::grpc::ErrorCode::UNEXPECTED_ERROR},
+        {SERVER_UNSUPPORTED_ERROR, ::milvus::grpc::ErrorCode::UNEXPECTED_ERROR},
+        {SERVER_NULL_POINTER, ::milvus::grpc::ErrorCode::UNEXPECTED_ERROR},
+        {SERVER_INVALID_ARGUMENT, ::milvus::grpc::ErrorCode::ILLEGAL_ARGUMENT},
+        {SERVER_FILE_NOT_FOUND, ::milvus::grpc::ErrorCode::FILE_NOT_FOUND},
+        {SERVER_NOT_IMPLEMENT, ::milvus::grpc::ErrorCode::UNEXPECTED_ERROR},
+        {SERVER_CANNOT_CREATE_FOLDER, ::milvus::grpc::ErrorCode::CANNOT_CREATE_FOLDER},
+        {SERVER_CANNOT_CREATE_FILE, ::milvus::grpc::ErrorCode::CANNOT_CREATE_FILE},
+        {SERVER_CANNOT_DELETE_FOLDER, ::milvus::grpc::ErrorCode::CANNOT_DELETE_FOLDER},
+        {SERVER_CANNOT_DELETE_FILE, ::milvus::grpc::ErrorCode::CANNOT_DELETE_FILE},
+        {SERVER_TABLE_NOT_EXIST, ::milvus::grpc::ErrorCode::TABLE_NOT_EXISTS},
+        {SERVER_INVALID_TABLE_NAME, ::milvus::grpc::ErrorCode::ILLEGAL_TABLE_NAME},
+        {SERVER_INVALID_TABLE_DIMENSION, ::milvus::grpc::ErrorCode::ILLEGAL_DIMENSION},
+        {SERVER_INVALID_TIME_RANGE, ::milvus::grpc::ErrorCode::ILLEGAL_RANGE},
+        {SERVER_INVALID_VECTOR_DIMENSION, ::milvus::grpc::ErrorCode::ILLEGAL_DIMENSION},
+
+        {SERVER_INVALID_INDEX_TYPE, ::milvus::grpc::ErrorCode::ILLEGAL_INDEX_TYPE},
+        {SERVER_INVALID_ROWRECORD, ::milvus::grpc::ErrorCode::ILLEGAL_ROWRECORD},
+        {SERVER_INVALID_ROWRECORD_ARRAY, ::milvus::grpc::ErrorCode::ILLEGAL_ROWRECORD},
+        {SERVER_INVALID_TOPK, ::milvus::grpc::ErrorCode::ILLEGAL_TOPK},
+        {SERVER_INVALID_NPROBE, ::milvus::grpc::ErrorCode::ILLEGAL_ARGUMENT},
+        {SERVER_INVALID_INDEX_NLIST, ::milvus::grpc::ErrorCode::ILLEGAL_NLIST},
+        {SERVER_INVALID_INDEX_METRIC_TYPE, ::milvus::grpc::ErrorCode::ILLEGAL_METRIC_TYPE},
+        {SERVER_INVALID_INDEX_FILE_SIZE, ::milvus::grpc::ErrorCode::ILLEGAL_ARGUMENT},
+        {SERVER_ILLEGAL_VECTOR_ID, ::milvus::grpc::ErrorCode::ILLEGAL_VECTOR_ID},
+        {SERVER_ILLEGAL_SEARCH_RESULT, ::milvus::grpc::ErrorCode::ILLEGAL_SEARCH_RESULT},
+        {SERVER_CACHE_FULL, ::milvus::grpc::ErrorCode::CACHE_FAILED},
+        {DB_META_TRANSACTION_FAILED, ::milvus::grpc::ErrorCode::META_FAILED},
+        {SERVER_BUILD_INDEX_ERROR, ::milvus::grpc::ErrorCode::BUILD_INDEX_ERROR},
+        {SERVER_OUT_OF_MEMORY, ::milvus::grpc::ErrorCode::OUT_OF_MEMORY},
+    };
+
+    if (code_map.find(code) != code_map.end()) {
+        return code_map.at(code);
+    } else {
+        return ::milvus::grpc::ErrorCode::UNEXPECTED_ERROR;
+    }
+}
 
 GrpcRequestHandler::GrpcRequestHandler(const std::shared_ptr<opentracing::Tracer>& tracer)
     : tracer_(tracer), random_num_generator_() {
@@ -130,185 +152,334 @@ GrpcRequestHandler::random_id() const {
 ::grpc::Status
 GrpcRequestHandler::CreateTable(::grpc::ServerContext* context, const ::milvus::grpc::TableSchema* request,
                                 ::milvus::grpc::Status* response) {
-    BaseRequestPtr request_ptr = CreateTableRequest::Create(context_map_[context], request);
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
-    SET_TRACING_TAG(*response, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    Status status = request_handler_.CreateTable(context_map_[context], request->table_name(), request->dimension(),
+                                                 request->index_file_size(), request->metric_type());
+    SET_RESPONSE(response, status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::HasTable(::grpc::ServerContext* context, const ::milvus::grpc::TableName* request,
                              ::milvus::grpc::BoolReply* response) {
+    CHECK_NULLPTR_RETURN(request);
+
     bool has_table = false;
-    BaseRequestPtr request_ptr = HasTableRequest::Create(context_map_[context], request->table_name(), has_table);
-    ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+
+    Status status = request_handler_.HasTable(context_map_[context], request->table_name(), has_table);
     response->set_bool_reply(has_table);
-    SET_RESPONSE(response, grpc_status, context);
+    SET_RESPONSE(response->mutable_status(), status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::DropTable(::grpc::ServerContext* context, const ::milvus::grpc::TableName* request,
                               ::milvus::grpc::Status* response) {
-    BaseRequestPtr request_ptr = DropTableRequest::Create(context_map_[context], request->table_name());
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
-    SET_TRACING_TAG(*response, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    Status status = request_handler_.DropTable(context_map_[context], request->table_name());
+
+    SET_RESPONSE(response, status, context);
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::CreateIndex(::grpc::ServerContext* context, const ::milvus::grpc::IndexParam* request,
                                 ::milvus::grpc::Status* response) {
-    BaseRequestPtr request_ptr = CreateIndexRequest::Create(context_map_[context], request);
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
-    SET_TRACING_TAG(*response, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    Status status = request_handler_.CreateIndex(context_map_[context], request->table_name(),
+                                                 request->index().index_type(), request->index().nlist());
+
+    SET_RESPONSE(response, status, context);
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::Insert(::grpc::ServerContext* context, const ::milvus::grpc::InsertParam* request,
                            ::milvus::grpc::VectorIds* response) {
-    BaseRequestPtr request_ptr = InsertRequest::Create(context_map_[context], request, response);
-    ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
-    SET_RESPONSE(response, grpc_status, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    int64_t record_data_size = 0;
+    for (auto& record : request->row_record_array()) {
+        record_data_size += record.vector_data_size();
+    }
+
+    std::vector<float> record_array(record_data_size, 0.0f);
+    int64_t offset = 0;
+    for (auto& record : request->row_record_array()) {
+        memcpy(&record_array[offset], record.vector_data().data(), record.vector_data_size() * sizeof(float));
+        offset += record.vector_data_size();
+    }
+
+    std::vector<int64_t> id_array;
+    if (request->row_id_array_size() > 0) {
+        id_array.resize(request->row_id_array_size());
+        memcpy(id_array.data(), request->row_id_array().data(), request->row_id_array_size() * sizeof(int64_t));
+    }
+
+    Status status =
+        request_handler_.Insert(context_map_[context], request->table_name(), request->row_record_array_size(),
+                                record_array, request->partition_tag(), id_array);
+
+    response->mutable_vector_id_array()->Resize(static_cast<int>(id_array.size()), 0);
+    memcpy(response->mutable_vector_id_array()->mutable_data(), id_array.data(), id_array.size() * sizeof(int64_t));
+
+    SET_RESPONSE(response->mutable_status(), status, context);
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::Search(::grpc::ServerContext* context, const ::milvus::grpc::SearchParam* request,
                            ::milvus::grpc::TopKQueryResult* response) {
-    std::vector<std::string> file_id_array;
-    BaseRequestPtr request_ptr = SearchRequest::Create(context_map_[context], request, file_id_array, response);
-    ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
-    SET_RESPONSE(response, grpc_status, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    int64_t record_data_size = 0;
+    for (auto& record : request->query_record_array()) {
+        record_data_size += record.vector_data_size();
+    }
+
+    std::vector<float> record_array(record_data_size);
+    int64_t offset = 0;
+    for (auto& record : request->query_record_array()) {
+        memcpy(&record_array[offset], record.vector_data().data(), record.vector_data_size() * sizeof(float));
+        offset += record.vector_data_size();
+    }
+
+    std::vector<Range> ranges;
+    for (auto& range : request->query_range_array()) {
+        ranges.emplace_back(range.start_value(), range.end_value());
+    }
+
+    std::vector<std::string> partitions;
+    for (auto& partition : request->partition_tag_array()) {
+        partitions.emplace_back(partition);
+    }
+
+    std::vector<std::string> file_ids;
+    TopKQueryResult result;
+
+    Status status =
+        request_handler_.Search(context_map_[context], request->table_name(), request->query_record_array_size(),
+                                record_array, ranges, request->topk(), request->nprobe(), partitions, file_ids, result);
+
+    // construct result
+    response->set_row_num(result.row_num_);
+
+    response->mutable_ids()->Resize(static_cast<int>(result.id_list_.size()), 0);
+    memcpy(response->mutable_ids()->mutable_data(), result.id_list_.data(), result.id_list_.size() * sizeof(int64_t));
+
+    response->mutable_distances()->Resize(static_cast<int>(result.distance_list_.size()), 0.0);
+    memcpy(response->mutable_distances()->mutable_data(), result.distance_list_.data(),
+           result.distance_list_.size() * sizeof(float));
+
+    SET_RESPONSE(response->mutable_status(), status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::SearchInFiles(::grpc::ServerContext* context, const ::milvus::grpc::SearchInFilesParam* request,
                                   ::milvus::grpc::TopKQueryResult* response) {
-    std::vector<std::string> file_id_array;
-    for (int i = 0; i < request->file_id_array_size(); i++) {
-        file_id_array.push_back(request->file_id_array(i));
+    CHECK_NULLPTR_RETURN(request);
+
+    std::vector<std::string> file_ids;
+    for (auto& file_id : request->file_id_array()) {
+        file_ids.emplace_back(file_id);
     }
-    ::milvus::grpc::SearchInFilesParam* request_mutable = const_cast<::milvus::grpc::SearchInFilesParam*>(request);
-    BaseRequestPtr request_ptr =
-        SearchRequest::Create(context_map_[context], request_mutable->mutable_search_param(), file_id_array, response);
-    ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
-    SET_RESPONSE(response, grpc_status, context);
+
+    auto* search_request = &request->search_param();
+
+    int64_t record_data_size = 0;
+    for (auto& record : search_request->query_record_array()) {
+        record_data_size += record.vector_data_size();
+    }
+
+    std::vector<float> record_array(record_data_size);
+    int64_t offset = 0;
+    for (auto& record : search_request->query_record_array()) {
+        memcpy(&record_array[offset], record.vector_data().data(), record.vector_data_size() * sizeof(float));
+        offset += record.vector_data_size();
+    }
+
+    std::vector<Range> ranges;
+    for (auto& range : search_request->query_range_array()) {
+        ranges.emplace_back(range.start_value(), range.end_value());
+    }
+
+    std::vector<std::string> partitions;
+    for (auto& partition : search_request->partition_tag_array()) {
+        partitions.emplace_back(partition);
+    }
+
+    TopKQueryResult result;
+
+    Status status = request_handler_.Search(
+        context_map_[context], search_request->table_name(), search_request->query_record_array_size(), record_array,
+        ranges, search_request->topk(), search_request->nprobe(), partitions, file_ids, result);
+
+    // construct result
+    response->set_row_num(result.row_num_);
+
+    response->mutable_ids()->Resize(static_cast<int>(result.id_list_.size()), 0);
+    memcpy(response->mutable_ids()->mutable_data(), result.id_list_.data(), result.id_list_.size() * sizeof(int64_t));
+
+    response->mutable_distances()->Resize(static_cast<int>(result.distance_list_.size()), 0.0);
+    memcpy(response->mutable_distances()->mutable_data(), result.distance_list_.data(),
+           result.distance_list_.size() * sizeof(float));
+    SET_RESPONSE(response->mutable_status(), status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::DescribeTable(::grpc::ServerContext* context, const ::milvus::grpc::TableName* request,
                                   ::milvus::grpc::TableSchema* response) {
-    BaseRequestPtr request_ptr = DescribeTableRequest::Create(context_map_[context], request->table_name(), response);
-    ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
-    SET_RESPONSE(response, grpc_status, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    TableSchema table_schema;
+    Status status = request_handler_.DescribeTable(context_map_[context], request->table_name(), table_schema);
+    response->set_table_name(table_schema.table_name_);
+    response->set_dimension(table_schema.dimension_);
+    response->set_index_file_size(table_schema.index_file_size_);
+    response->set_metric_type(table_schema.metric_type_);
+
+    SET_RESPONSE(response->mutable_status(), status, context);
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::CountTable(::grpc::ServerContext* context, const ::milvus::grpc::TableName* request,
                                ::milvus::grpc::TableRowCount* response) {
+    CHECK_NULLPTR_RETURN(request);
+
     int64_t row_count = 0;
-    BaseRequestPtr request_ptr = CountTableRequest::Create(context_map_[context], request->table_name(), row_count);
-    ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    Status status = request_handler_.CountTable(context_map_[context], request->table_name(), row_count);
     response->set_table_row_count(row_count);
-    SET_RESPONSE(response, grpc_status, context);
+    SET_RESPONSE(response->mutable_status(), status, context);
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::ShowTables(::grpc::ServerContext* context, const ::milvus::grpc::Command* request,
                                ::milvus::grpc::TableNameList* response) {
-    BaseRequestPtr request_ptr = ShowTablesRequest::Create(context_map_[context], response);
-    ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
-    SET_RESPONSE(response, grpc_status, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    std::vector<std::string> tables;
+    Status status = request_handler_.ShowTables(context_map_[context], tables);
+    for (auto& table : tables) {
+        response->add_table_names(table);
+    }
+    SET_RESPONSE(response->mutable_status(), status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::Cmd(::grpc::ServerContext* context, const ::milvus::grpc::Command* request,
                         ::milvus::grpc::StringReply* response) {
-    std::string result;
-    BaseRequestPtr request_ptr = CmdRequest::Create(context_map_[context], request->cmd(), result);
-    ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
-    response->set_string_reply(result);
-    SET_RESPONSE(response, grpc_status, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    std::string reply;
+    Status status = request_handler_.Cmd(context_map_[context], request->cmd(), reply);
+    response->set_string_reply(reply);
+    SET_RESPONSE(response->mutable_status(), status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::DeleteByDate(::grpc::ServerContext* context, const ::milvus::grpc::DeleteByDateParam* request,
                                  ::milvus::grpc::Status* response) {
-    BaseRequestPtr request_ptr = DeleteByDateRequest::Create(context_map_[context], request);
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
-    SET_TRACING_TAG(*response, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    Range range(request->range().start_value(), request->range().end_value());
+    Status status = request_handler_.DeleteByRange(context_map_[context], request->table_name(), range);
+    SET_RESPONSE(response, status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::PreloadTable(::grpc::ServerContext* context, const ::milvus::grpc::TableName* request,
                                  ::milvus::grpc::Status* response) {
-    BaseRequestPtr request_ptr = PreloadTableRequest::Create(context_map_[context], request->table_name());
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
-    SET_TRACING_TAG(*response, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    Status status = request_handler_.PreloadTable(context_map_[context], request->table_name());
+    SET_RESPONSE(response, status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::DescribeIndex(::grpc::ServerContext* context, const ::milvus::grpc::TableName* request,
                                   ::milvus::grpc::IndexParam* response) {
-    BaseRequestPtr request_ptr = DescribeIndexRequest::Create(context_map_[context], request->table_name(), response);
-    ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
-    SET_RESPONSE(response, grpc_status, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    IndexParam param;
+    Status status = request_handler_.DescribeIndex(context_map_[context], request->table_name(), param);
+    response->set_table_name(param.table_name_);
+    response->mutable_index()->set_index_type(param.index_type_);
+    response->mutable_index()->set_nlist(param.nlist_);
+    SET_RESPONSE(response->mutable_status(), status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::DropIndex(::grpc::ServerContext* context, const ::milvus::grpc::TableName* request,
                               ::milvus::grpc::Status* response) {
-    BaseRequestPtr request_ptr = DropIndexRequest::Create(context_map_[context], request->table_name());
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
-    SET_TRACING_TAG(*response, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    Status status = request_handler_.DropIndex(context_map_[context], request->table_name());
+    SET_RESPONSE(response, status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::CreatePartition(::grpc::ServerContext* context, const ::milvus::grpc::PartitionParam* request,
                                     ::milvus::grpc::Status* response) {
-    BaseRequestPtr request_ptr = CreatePartitionRequest::Create(context_map_[context], request);
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
-    SET_TRACING_TAG(*response, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    Status status = request_handler_.CreatePartition(context_map_[context], request->table_name(),
+                                                     request->partition_name(), request->tag());
+    SET_RESPONSE(response, status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::ShowPartitions(::grpc::ServerContext* context, const ::milvus::grpc::TableName* request,
                                    ::milvus::grpc::PartitionList* response) {
-    BaseRequestPtr request_ptr = ShowPartitionsRequest::Create(context_map_[context], request->table_name(), response);
-    ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
-    SET_RESPONSE(response, grpc_status, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    std::vector<PartitionParam> partitions;
+    Status status = request_handler_.ShowPartitions(context_map_[context], request->table_name(), partitions);
+    for (auto& partition : partitions) {
+        milvus::grpc::PartitionParam* param = response->add_partition_array();
+        param->set_table_name(partition.table_name_);
+        param->set_partition_name(partition.partition_name_);
+        param->set_tag(partition.tag_);
+    }
+
+    SET_RESPONSE(response->mutable_status(), status, context);
+
     return ::grpc::Status::OK;
 }
 
 ::grpc::Status
 GrpcRequestHandler::DropPartition(::grpc::ServerContext* context, const ::milvus::grpc::PartitionParam* request,
                                   ::milvus::grpc::Status* response) {
-    BaseRequestPtr request_ptr = DropPartitionRequest::Create(context_map_[context], request);
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
-    SET_TRACING_TAG(*response, context);
+    CHECK_NULLPTR_RETURN(request);
+
+    Status status = request_handler_.DropPartition(context_map_[context], request->table_name(),
+                                                   request->partition_name(), request->tag());
+    SET_RESPONSE(response, status, context);
+
     return ::grpc::Status::OK;
 }